requirements.txt
playwright
requests
fake-useragent
python-dotenv
